# Improved concurrency for multi-threaded indexing