Concurrent tree building is faster on my 12 core laptop for ~130M point clouds, in this arrangement.

The intuition is to avoid having threads wait for both the left and right subtrees if they are both asynchronous.
In this arrangment we first recurse to left on the same thread (depth-first to the left) with an async recursion of the right subtree concurrently.
Once the left is all visited we'll recurse to the right with the current thread.

What this is aiming to avoid is having too many threads blocked (doing no work) waiting for async processing to complete on both left and right.

I see around a 2x speedup for 12 cores and ~130M, around 30 seconds rather than 60 seconds.