@jlblancoc 2024-03-05
Great contribution! Let me check this against valgrind and do some minimal benchmarking before merging.