@jlblancoc 2024-03-11
I updated [nanoflann benchmark](https://github.com/MRPT/nanoflann-benchmark) to make some quick benchmarking with this PR. The results are good but, as usual, the optimal approach depends on the data distribution and dataset size: 

Build index profiling for the KITTI 04 sequence pointclouds:

| Method | 1 thread | 4 threads | 8 threads | auto # threads (=0) | 
|---|---|---|---|---|
Baseline | 8.993637 ms | 6.922916 ms | **6.759240 ms** | **7.775914 ms** |
This PR | 8.999447 ms | **6.235389 ms** | 7.164032 ms | 8.716205 ms |

Some graphs:

**For 4 threads vs 1 thread**

![Screenshot from 2024-03-12 00-04-33](https://github.com/jlblancoc/nanoflann/assets/5497818/05213811-ad0a-4fb5-aa66-77da171189fe)

**For 8 threads vs 1 thread**

![Screenshot from 2024-03-12 00-03-34](https://github.com/jlblancoc/nanoflann/assets/5497818/988ad8c5-addb-415a-9306-84c595252682)


The advantage of this PR is clearer for 4 threads than for 8 threads... for this particular pointcloud. 

Anyway, I feel positive about the **logic** of the change, and valgrind (including helgrind) is happy, so I'm merging it... 
Thanks!