@nigels-com 2024-03-11
Interesting results.  Modest speedup for this point cloud.
Do you know how many points there are?
Our typical workload is 100M to 1B points.