@nigels-com 2024-03-12
I'm a bit intrigued about four threads being faster than eight threads.
I'd like to dig into that a little more.