@jlblancoc 2024-03-13
@nigels-com The exact place to force a fixed number of threads in our benchmark is here: 
https://github.com/MRPT/nanoflann-benchmark/blob/master/benchmarkTool/realTests/benchmark_nanoflann_real.cpp#L169-L177

feel free of using that in your project/dataset just to see what happens... 

My guess is that results may depend a lot depending on how "random" the distribution of consecutive points are. In 3D LIDARs (as KITTI), point coordinates have a strong correlation with their neighbors in the sequence.