@nigels-com 2024-03-13
Thanks for the details.

Our points are probably less coherent because the Hovermap lidar is both spinning and moving through space.
We do see better performance sorting them upstream of nanoflann.
So my number and graphs are likely to be quite different to this benchmark.

https://emesent.com/hovermap-series/